+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## vNext (TBD)
 
 ### Enhancements
-* None
+* Added `Text` operator support. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
+* Added `Phrase` operator support. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
+* Added some geojson types. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
 
 ### Fixed
 * None